+++ src/UniFiDevice.js
@@ -28,10 +28,6 @@ module.exports = class UniFiDevice {
     return this.homeKitAccessory.context.device_id;
   }
 
-  get port_overrides() {
-    return this.homeKitAccessory.context.port_overrides;
-  }
-
   get port_onMode() {
     return this.homeKitAccessory.context.port_onMode;
   }
@@ -40,38 +36,39 @@ module.exports = class UniFiDevice {
     return this.homeKitAccessory.context.port_idx;
   }
 
-  matches(device, port, devicePortConfig) {
+  matches(device, devicePortConfig) {
     return (
       this.mac === device.mac &&
-      this.port_idx === port.port_idx &&
+      this.port_idx === devicePortConfig.idx &&
       this.port_onMode === devicePortConfig.onMode
     );
   }
 
-  static getContextForDevicePort(site, device, port, devicePortConfig) {
+  static getContextForDevicePort(site, device, devicePortConfig) {
     return {
       site,
       mac: device.mac,
       device_id: device.device_id,
-      port_overrides: device.port_overrides,
       port_onMode: devicePortConfig.onMode,
-      port_idx: port.port_idx
+      port_idx: devicePortConfig.idx
     };
   }
 
   async update(site, device, port, devicePortConfig) {
-    this.homeKitAccessory.context = UniFiDevice.getContextForDevicePort(site, device, port, devicePortConfig);
+    this.homeKitAccessory.context = UniFiDevice.getContextForDevicePort(site, device, devicePortConfig);
 
     this.homeKitAccessory.getService(Service.AccessoryInformation)
-      .setCharacteristic(Characteristic.Name, devicePortConfig.name || port.name || device.name + ' - Port ' + port.port_idx)
+      .setCharacteristic(Characteristic.Name, devicePortConfig.name)
       .setCharacteristic(Characteristic.Manufacturer, 'Stefan Kienzle')
       .setCharacteristic(Characteristic.Model, device.name || device.model)
-      .setCharacteristic(Characteristic.SerialNumber, device.mac + '-' + port.port_idx + (devicePortConfig.onMode === 'power_cycle' ? '-pc' : ''));
+      .setCharacteristic(Characteristic.SerialNumber, devicePortConfig.mac + '-' + devicePortConfig.idx + (devicePortConfig.onMode === 'power_cycle' ? '-pc' : ''));
+
+    if (this.port_onMode === 'power_cycle') {
+      this.getCharacteristic(Characteristic.On).updateValue(false);
+    }
 
     if (!this.changePending) {
-      if (this.port_onMode === 'power_cycle') {
-        this.getCharacteristic(Characteristic.On).updateValue(false);
-      } else {
+      if (this.port_onMode !== 'power_cycle') {
         this.getCharacteristic(Characteristic.On).updateValue(port.poe_mode !== 'off');
       }
     }
@@ -91,20 +88,46 @@ module.exports = class UniFiDevice {
     return this.getService().getCharacteristic(characteristic);
   }
 
-  _setAllProperties() {
+  async _setAllProperties() {
+    let port_override = {};
+    let port = {};
+    let devices, foundDevice;
+
+    // Refresh devices to get current port_overrides
+    try {
+      devices = await this.plugin.client.getDevices(this.site.name);
+      foundDevice = devices.data.find(device => device.mac === this.mac);
+    } catch (e) {};
+
+    if (!foundDevice) {
+      this.plugin.log.warn(`Device (ID: ${this.device_id}, MAC: ${this.mac}) doesn't exists`);
+      this.changePending = false;
+      return await this.plugin.refreshDevices();
+    }
+
+    port = foundDevice.port_table.find(port => port.port_idx === this.port_idx, this);
+
+    if (Array.isArray(foundDevice.port_overrides)) {
+      port_override = foundDevice.port_overrides.find(port_override => port_override.port_idx === this.port_idx, this);
+    }
+
     // Just power cycle the port - no change to port config
     if (this.port_onMode === 'power_cycle') {
-      let poePortEnabled = false;
+      let poePortEnabled;
 
-      for (let override of this.port_overrides) {
-        if (override.port_idx === this.port_idx) {
-          poePortEnabled = override.poe_mode !== 'off';
-        }
+      if (port_override) {
+        poePortEnabled = port_override.poe_mode !== 'off';
+      } else {
+        poePortEnabled = port.poe_mode !== 'off';
       }
 
       if (!poePortEnabled) {
-        this.plugin.log.info(`Device ${this.device_id}: Power Cycle not available - POE is turned off`);
+        this.plugin.log.info(`Device (ID: ${this.device_id}, MAC: ${this.mac}): Power Cycle not available - POE is turned off`);
         this.changePending = false;
+        
+        // Reset characteristic value - is always "false"
+        this.getCharacteristic(Characteristic.On).updateValue(false);
+
         return;
       }
 
@@ -120,14 +143,20 @@ module.exports = class UniFiDevice {
 
     // Change port config to new poe_mode value
     } else {
-      for (let override of this.port_overrides) {
-        if (override.port_idx === this.port_idx) {
-          override.poe_mode = this.getCharacteristic(Characteristic.On).value ? this.port_onMode : 'off'
-        }
+      let port_overrides = foundDevice.port_overrides || [];
+
+      if (port_override) {
+        port_override.poe_mode = this.getCharacteristic(Characteristic.On).value ? this.port_onMode : 'off';
+      } else {
+        port_overrides.push({
+          port_idx: port.port_idx,
+          portconf_id: port.portconf_id,
+          poe_mode: this.getCharacteristic(Characteristic.On).value ? this.port_onMode : 'off'
+        });
       }
 
       let properties = {
-        port_overrides: this.port_overrides,
+        port_overrides: port_overrides,
         device_id: this.device_id
       };
 
@@ -138,7 +167,7 @@ module.exports = class UniFiDevice {
   }
 
   async setProperties(properties) {
-    this.plugin.log.info(`Device ${this.device_id}: Setting properties: ${JSON.stringify(properties)}`);
+    this.plugin.log.info(`Device (ID: ${this.device_id}, MAC: ${this.mac}): Update "port_overrides"`);
 
     try {
       await this.plugin.client.setDevice(this.site.name, this.device_id, properties);
@@ -151,7 +180,10 @@ module.exports = class UniFiDevice {
   }
 
   async setPowerCycle(properties) {
-    this.plugin.log.info(`Device ${this.device_id}: Power Cylce: ${JSON.stringify(properties)}`);
+    this.plugin.log.info(`Device (ID: ${this.device_id}, MAC: ${this.mac}): Command "power_cycle"`);
+    
+    // Reset characteristic value - is always "false"
+    this.getCharacteristic(Characteristic.On).updateValue(false);
 
     try {
       await this.plugin.client.setPowerCycle(this.site.name, properties);